Fix SpriteViewTest bug.
SpriteViewTest had a hardcoded reference to the "pacman" sprite, which
was removed and replaced with the 3 separate sprites "pacman_open",
"pacman_halfopen", "pacman_closed" in 9c8278e in preparation for
animation.

The "pacman" costume lookup was silently failing and returning a
half-scale orange box, which was breaking the SpriteView test.